调整repo
