+++ pkg/controller/controller.go
@@ -23,7 +23,6 @@ import (
 	"github.com/caoyingjunz/pixiu/pkg/controller/audit"
 	"github.com/caoyingjunz/pixiu/pkg/controller/auth"
 	"github.com/caoyingjunz/pixiu/pkg/controller/cluster"
-	"github.com/caoyingjunz/pixiu/pkg/controller/helm"
 	"github.com/caoyingjunz/pixiu/pkg/controller/plan"
 	"github.com/caoyingjunz/pixiu/pkg/controller/tenant"
 	"github.com/caoyingjunz/pixiu/pkg/controller/user"
@@ -37,7 +36,6 @@ type PixiuInterface interface {
 	plan.PlanGetter
 	audit.AuditGetter
 	auth.AuthGetter
-	helm.RepoGetter
 }
 
 type pixiu struct {
@@ -46,13 +44,12 @@ type pixiu struct {
 	enforcer *casbin.SyncedEnforcer
 }
 
-func (p *pixiu) Cluster() cluster.Interface   { return cluster.NewCluster(p.cc, p.factory, p.enforcer) }
-func (p *pixiu) Tenant() tenant.Interface     { return tenant.NewTenant(p.cc, p.factory) }
-func (p *pixiu) User() user.Interface         { return user.NewUser(p.cc, p.factory, p.enforcer) }
-func (p *pixiu) Plan() plan.Interface         { return plan.NewPlan(p.cc, p.factory) }
-func (p *pixiu) Audit() audit.Interface       { return audit.NewAudit(p.cc, p.factory) }
-func (p *pixiu) Auth() auth.Interface         { return auth.NewAuth(p.factory, p.enforcer) }
-func (p *pixiu) Repositories() helm.Interface { return helm.NewRepositories(p.factory) }
+func (p *pixiu) Cluster() cluster.Interface { return cluster.NewCluster(p.cc, p.factory, p.enforcer) }
+func (p *pixiu) Tenant() tenant.Interface   { return tenant.NewTenant(p.cc, p.factory) }
+func (p *pixiu) User() user.Interface       { return user.NewUser(p.cc, p.factory, p.enforcer) }
+func (p *pixiu) Plan() plan.Interface       { return plan.NewPlan(p.cc, p.factory) }
+func (p *pixiu) Audit() audit.Interface     { return audit.NewAudit(p.cc, p.factory) }
+func (p *pixiu) Auth() auth.Interface       { return auth.NewAuth(p.factory, p.enforcer) }
 func New(cfg config.Config, f db.ShareDaoFactory, enforcer *casbin.SyncedEnforcer) PixiuInterface {
 	return &pixiu{
 		cc:       cfg,
